+++ libutopia2/utopia2/hashmap.h
@@ -1,7 +1,7 @@
 /*****************************************************************************
  *  
  *   This file is part of the Utopia Documents application.
- *       Copyright (c) 2008-2014 Lost Island Labs
+ *       Copyright (c) 2008-2016 Lost Island Labs
  *           <info@utopiadocs.com>
  *   
  *   Utopia Documents is free software: you can redistribute it and/or modify
@@ -62,9 +62,9 @@ namespace Utopia
     {
     public:
         // Hash function
-        unsigned long operator () (ConcreteKeyType* key_)
+        unsigned long long operator () (ConcreteKeyType* key_)
         {
-            return (unsigned long) key_ >> 3;
+            return (unsigned long long) key_ >> 3;
         }
     };
